afxdp/types/shared_cos_lease: relocate inline tests to shared_cos_lease_tests.rs
Same playbook. 71-LOC inline test block extracted. shared_cos_lease.rs: 522 → 454 LOC.
cargo build/test 92/878.
