+++ userspace-dp/src/afxdp/types/shared_cos_lease_tests.rs
@@ -1,0 +1,68 @@
+// Tests for afxdp/types/shared_cos_lease.rs — relocated from inline
+// `#[cfg(test)] mod tests` to keep shared_cos_lease.rs under the modularity-discipline
+// LOC threshold. Loaded as a sibling submodule via
+// `#[path = "shared_cos_lease_tests.rs"]` from shared_cos_lease.rs.
+
+use super::*;
+use std::mem::align_of;
+
+// #694 / #711: `FlowRrRing` invariant pins.
+//
+// The ring is the SFQ round-robin cursor storage. Every bug class
+// that can break it is pinned here so a future refactor that
+// changes the indexing math, the wrap condition, or the head/len
+// update order fails loudly in CI instead of during live
+// validation.
+
+fn shared_cos_lease_snapshot(lease: &SharedCoSRootLease) -> (u64, u64, u64) {
+    let (available_tokens, outstanding_leased_tokens) =
+        unpack_shared_cos_lease_credits(lease.state.credits.load(Ordering::Relaxed));
+    let last_refill_ns = lease.state.last_refill_ns.load(Ordering::Relaxed);
+    (available_tokens, outstanding_leased_tokens, last_refill_ns)
+}
+
+#[test]
+fn shared_cos_root_lease_refill_respects_outstanding_burst_credit() {
+    let lease = SharedCoSRootLease::new(10_000_000, 16_000, 1);
+    lease
+        .state
+        .credits
+        .store(pack_shared_cos_lease_credits(0, 4_000), Ordering::Relaxed);
+    lease.state.last_refill_ns.store(1, Ordering::Relaxed);
+
+    refill_shared_cos_lease_state(lease.config, &lease.state, 1_000_000_001);
+
+    let (available_tokens, outstanding_leased_tokens, _) = shared_cos_lease_snapshot(&lease);
+    assert_eq!(
+        available_tokens,
+        lease.config.burst_bytes - outstanding_leased_tokens
+    );
+}
+
+#[test]
+fn shared_cos_root_lease_release_unused_preserves_total_burst_bound() {
+    let lease = SharedCoSRootLease::new(10_000_000, 16_000, 1);
+    lease.state.credits.store(
+        pack_shared_cos_lease_credits(lease.config.burst_bytes, 4_000),
+        Ordering::Relaxed,
+    );
+
+    lease.release_unused(1_500);
+
+    let (available_tokens, outstanding_leased_tokens, _) = shared_cos_lease_snapshot(&lease);
+    assert_eq!(
+        available_tokens + outstanding_leased_tokens,
+        lease.config.burst_bytes
+    );
+}
+
+#[test]
+fn shared_cos_lease_state_is_cacheline_aligned() {
+    assert_eq!(align_of::<SharedCoSLeaseState>(), 64);
+}
+
+#[test]
+fn shared_cos_lease_config_clamps_burst_to_packed_range() {
+    let lease = SharedCoSRootLease::new(10_000_000, u64::MAX, 1);
+    assert_eq!(lease.config.burst_bytes, u32::MAX as u64);
+}
